lib/js/datachannel_test_2peer/send/013.html

<!DOCTYPE html>
<html lang="en" class="a">
	<head>
		<meta charset="utf-8">
		<!-- Testharness Timeout to long -> 60 Seconds-->
		<meta name="timeout" content="long">
		<title>WebRTC - Send Test</title>
		<script src="../resources/testharness.js"></script>
		<script src="../resources/testharnessreport.js"></script><link href="../resources/testharness.css" type="text/css" rel="stylesheet"/>
		<!-- Including the Firebase.js to send the ICE Candidate to the other peer-->
		<script type="text/javascript" src='https://cdn.firebase.com/v0/firebase.js'></script>
	</head>
	<body class="a">
		<div>
			<div id="signallingInfo"></div>
			
				
			
		</div>
		<hr />

		<!-- Testharness log -->
		<div id="log"></div>
		<script src="../datachannel2peers.js" type="text/javascript"></script>
		<script type="text/javascript">
			setup({
			    explicit_timeout : true
			});
            var label = "test";
            var dataChannel = null;
            var expected = "s", size = 25, partialLength = 0;
            for (var i = 0; i < size; i++) {
                expected += expected;
            }
            var test = async_test("Set up a DataChannel - send a message of size " + expected.length / 1024 / 1024 + " MB ", {timeout: 1200000});
            if (offerer) {
                test.step(function() {
                    dataChannel = peerConnection.createDataChannel(label);
                    dataChannel.onopen = test.step_func(function() {
                        try{
                    		dataChannel.send(expected);
                    	}catch(e){
                    		assert_unreached(e.name + ": " + e.message);
                    	} 
                    	var iV = setInterval(test.step_func(function(){
                    		console.log(dataChannel.bufferedAmount);
                    		if(dataChannel.readyState == "closed"){
                    			assert_unreached("The dataChannel was closed unexpectly!");
                    		}
                    		if(dataChannel.bufferedAmount == 0){
                    			clearInterval(iV);
                    			test.done();
                    		}
                    	}), 500);                                   
                    });
                });
            } else {
                peerConnection.ondatachannel = test.step_func(function(e) {
                    dataChannel = e.channel;
                    dataChannel.onmessage = test.step_func(function(e) {
                    	if ((e.data).length + partialLength == expected.length){
							assert_equals(e.data.length + partialLength, expected.length, "Received data not correct ");
                    		test.done();     
                        }else{
                        	partialLength += (e.data).length;
                        	console.log("Partial receive " + partialLength);
                        }
                    });
                });
            }
		</script>
	</body>
</html>

